Add files via upload
Added a new attempt to the project, working alone and using a ESP32 WROOM 32 instead of an Arduino UNO R3 board, this commit includes the use of the OLED SPI Display, the logic behind the start of the cycle and the end of it, as well as the complete updated pinout. The main difference with the other codes that were already on this repo, is the use of finite state machines to avoid the use of the delay function as it hoards the processor's activities.